refactored several files and created new ones for abstraction. Deleted unused files and added comments to important code
