--- frontend/src/renderer/Canvas/Canvas.jsx
@@ -1,13 +1,0 @@
-import { DndContext } from '@dnd-kit/core';
-import { PlayBlock, SampleBlock } from 'renderer/Blocks/draggableBlocks';
-
-// import { Draggable } from '../Components/draggable';
-
-export default function Canvas() {
-  return (
-    <DndContext>
-      <PlayBlock />
-      <PlayBlock />
-    </DndContext>
-  );
-}
