--- frontend/src/renderer/Components/Knob.jsx
@@ -1,24 +1,0 @@
-// import React from 'react';
-// import { Knob } from 'primereact/knob';
-
-// export default function KnobComponent() {
-//   const [value, setValue] = React.useState(0);
-
-//   const onChange = (e) => {
-//     setValue(e.value);
-//   };
-
-//   return (
-//     <div>
-//       <Knob
-//         value={value}
-//         onChange={onChange}
-//         valueColor={'Turquoise'}
-//         rangeColor={'White'}
-//         rangeWidth={10}
-//         // size={50}
-//       />
-//       {value}
-//     </div>
-//   );
-// }
