+++ frontend/src/renderer/Menu/BlockDeleteButton.jsx
