+++ frontend/src/renderer/Menu/LoopDeleteButton.jsx
