+++ frontend/src/renderer/styles/Droppable.css
