+++ frontend/src/renderer/styles/Item.css
@@ -1,0 +1,8 @@
+.item {
+    margin-bottom: 5px;
+    padding-left: 5px;
+    border-radius: 5px;
+    user-select: none;  
+    background-color: rgb(115, 195, 240);
+  }
+
